Run CI against multiple DB adapters, Ruby and Rails versions (#16)

+++ spec/support/from_scratch.rb
@@ -1,8 +1,20 @@
 # frozen_string_literal: true
 
 RSpec.configure do |config|
+  config.before(:suite) do
+    next unless TestHelpers.solid_adapter?
+
+    db_adapter = TestHelpers.adapter_env.delete_prefix("solid/").to_sym
+    SolidTestHelpers.setup_database(db_adapter)
+  end
+
   config.before do
     TestHelpers.cache.clear
-    TestHelpers.redis.flushdb
+
+    if TestHelpers.solid_adapter?
+      SolidTestHelpers.clear_database
+    else
+      TestHelpers.redis.flushdb
+    end
   end
 end
